backend/requirements.txt
requests>=2.28.0
pyttsx3>=2.90
python-dotenv>=0.19.0